Move and update tests with new mock sources.
