+++ tests/danger/licenseAuditor.test.ts
@@ -231,52 +231,4 @@ describe("licenseAuditor DangerJS plugin module", () => {
       expect(global.fail).toBeCalledWith(message);
     });
   });
-
-  describe("licenseAuditor when there are no licenses", () => {
-    beforeAll(() => {
-      // mock the danger functions warn, fail, and markdown and attach them to the global object
-      global.warn = jest.fn();
-      global.fail = jest.fn();
-      global.markdown = jest.fn();
-    });
-
-    afterAll(() => {
-      // set the global object back to its original state
-      global.warn = undefined;
-      global.fail = undefined;
-      global.markdown = undefined;
-
-      // reset the mock for parseLicenseFactory
-      jest.unmock("../../src/auditor/parseLicenses");
-
-      // reset the licenseAuditor function
-      licenseAuditor = undefined;
-
-      // reset modules
-      jest.resetModules();
-    });
-
-    it("should call warn if there are no licenses", async () => {
-      // arrange
-      jest.doMock("../../src/auditor", () => {
-        // mock the findAllLicenses function to return an empty array
-        return jest.fn().mockImplementation(() => {
-          return [];
-        });
-      });
-      licenseAuditor = require("../../src/danger").licenseAuditor;
-      const config = {
-        failOnBlacklistedLicense: false,
-        projectPath: process.cwd(),
-        showMarkdownSummary: false,
-      };
-      const message = `There are no licenses to check.`;
-
-      // act
-      await licenseAuditor!(config);
-
-      // assert
-      expect(global.warn).toBeCalledWith(message);
-    });
-  });
 });
